+++ rxandroidble/src/main/java/com/polidea/rxandroidble/internal/connection/Connector.java
@@ -1,0 +1,10 @@
+package com.polidea.rxandroidble.internal.connection;
+
+
+import com.polidea.rxandroidble.RxBleConnection;
+import rx.Observable;
+
+public interface Connector {
+
+    Observable<RxBleConnection> prepareConnection(boolean autoConnect);
+}