+++ rxandroidble/src/main/java/com/polidea/rxandroidble/internal/operations/RxBleRadioOperationDisconnect.java
@@ -14,6 +14,7 @@ import com.polidea.rxandroidble.internal.RadioReleaseInterface;
 import com.polidea.rxandroidble.internal.RxBleLog;
 import com.polidea.rxandroidble.internal.RxBleRadioOperation;
 import com.polidea.rxandroidble.internal.connection.BluetoothGattProvider;
+import com.polidea.rxandroidble.internal.connection.ConnectionStateChangeListener;
 import com.polidea.rxandroidble.internal.connection.RxBleGattCallback;
 
 import javax.inject.Inject;
@@ -27,6 +28,8 @@ import rx.functions.Action0;
 import rx.functions.Action1;
 import rx.functions.Func1;
 
+import static com.polidea.rxandroidble.RxBleConnection.RxBleConnectionState.DISCONNECTED;
+import static com.polidea.rxandroidble.RxBleConnection.RxBleConnectionState.DISCONNECTING;
 import static rx.Observable.just;
 
 public class RxBleRadioOperationDisconnect extends RxBleRadioOperation<Void> {
@@ -37,6 +40,7 @@ public class RxBleRadioOperationDisconnect extends RxBleRadioOperation<Void> {
     private final BluetoothManager bluetoothManager;
     private final Scheduler mainThreadScheduler;
     private final TimeoutConfiguration timeoutConfiguration;
+    private final ConnectionStateChangeListener connectionStateChangeListener;
 
     @Inject
     RxBleRadioOperationDisconnect(
@@ -45,18 +49,20 @@ public class RxBleRadioOperationDisconnect extends RxBleRadioOperation<Void> {
             @Named(DeviceModule.MAC_ADDRESS) String macAddress,
             BluetoothManager bluetoothManager,
             @Named(ClientComponent.NamedSchedulers.MAIN_THREAD) Scheduler mainThreadScheduler,
-            @Named(DeviceModule.DISCONNECT_TIMEOUT) TimeoutConfiguration timeoutConfiguration) {
+            @Named(DeviceModule.DISCONNECT_TIMEOUT) TimeoutConfiguration timeoutConfiguration,
+            ConnectionStateChangeListener connectionStateChangeListener) {
         this.rxBleGattCallback = rxBleGattCallback;
         this.bluetoothGattProvider = bluetoothGattProvider;
         this.macAddress = macAddress;
         this.bluetoothManager = bluetoothManager;
         this.mainThreadScheduler = mainThreadScheduler;
         this.timeoutConfiguration = timeoutConfiguration;
+        this.connectionStateChangeListener = connectionStateChangeListener;
     }
 
     @Override
     protected void protectedRun(final Emitter<Void> emitter, RadioReleaseInterface radioReleaseInterface) {
-        //noinspection Convert2MethodRef
+        connectionStateChangeListener.onConnectionStateChange(DISCONNECTING);
         final BluetoothGatt bluetoothGatt = bluetoothGattProvider.getBluetoothGatt();
 
         if (bluetoothGatt == null) {
@@ -81,6 +87,7 @@ public class RxBleRadioOperationDisconnect extends RxBleRadioOperation<Void> {
                             new Action0() {
                                 @Override
                                 public void call() {
+                                    connectionStateChangeListener.onConnectionStateChange(DISCONNECTED);
                                     emitter.onCompleted();
                                 }
                             }
@@ -119,7 +126,7 @@ public class RxBleRadioOperationDisconnect extends RxBleRadioOperation<Void> {
                             .filter(new Func1<RxBleConnection.RxBleConnectionState, Boolean>() {
                                 @Override
                                 public Boolean call(RxBleConnection.RxBleConnectionState rxBleConnectionState) {
-                                    return rxBleConnectionState == RxBleConnection.RxBleConnectionState.DISCONNECTED;
+                                    return rxBleConnectionState == DISCONNECTED;
                                 }
                             })
                             .take(1)